BAH-4668. Moved LocationSelector, LocationContext, and LocationProvider from apps/home → packages/bahmni-widgets/src/location/
