--- apps/home/src/components/LocationSelector/index.ts
@@ -1,1 +1,0 @@
-export { LocationSelector } from './LocationSelector';
