+++ packages/bahmni-widgets/src/index.ts
@@ -59,3 +59,11 @@ export {
   resetWidgetRegistry,
   type WidgetConfig,
 } from './registry';
+
+// Location
+export {
+  LocationContext,
+  useLocation,
+  LocationProvider,
+  LocationSelector,
+} from './location';
