+++ packages/bahmni-widgets/src/location/__tests__/__mocks__/LocationSelectorMocks.ts
