+++ packages/bahmni-widgets/src/location/index.ts
@@ -1,3 +1,3 @@
 export { LocationContext, useLocation } from './LocationContext';
-export { type UserLocation } from '@bahmni/services';
 export { LocationProvider } from './LocationProvider';
+export { LocationSelector } from './LocationSelector';
