+++ packages/bahmni-widgets/src/location/styles/LocationSelector.module.scss
